Fix snake collision issue
There was an issue where the snake could turn back
into itself to go the opposite direction it was
going. So for instance, if it was going right, and
we went left, if there was any tail attached to
the snake, it would negate it and not count as a
collision like hitting the wall or the tail.
This was corrected. But now this introduced an
issue that if the game has just started,
the snake now wont go in the opposite direction
that it is moving. The snake should be allowed to
change direction freely when it hasn't grown yet,
as there's no risk of collision in that scenario.